Seems to be a bug in reading in attributes associated with skills. Going to correct whatevers wrong

+++ Dwarf/ECS/Systems/ai/BuildAi.cpp
@@ -52,11 +52,11 @@ void BuildAi::update(double duration)
 {
 	for (auto e : getEntities())
 	{
-		doBuild(e);
+		doBuild(e, duration);
 	}
 }
 
-void BuildAi::doBuild(const Entity & e)
+void BuildAi::doBuild(const Entity & e, const double & duration)
 {
 	WorkTemplate<BuilderTag> work;
 
@@ -223,58 +223,76 @@ void BuildAi::doBuild(const Entity & e)
 
 	else if (tag.step == BuilderTag::BUILD_BUILDING)
 	{
-		auto target = getBuilding(tag.buildingTarget.tag);
-		if (target == nullptr)
-			throw std::runtime_error("Building tag not found!");
-
-		// Add a method of incrementally increasing buildings completeness
-		// Should this use time, or use a random roll? Probably time based on difficulty and skill required
-
-		auto& buildingEntity = getWorld().getEntity(tag.buildingTarget.entity_id);
-
-		auto & building = buildingEntity.getComponent<Building>();
-
-		// Check every component for validity
-		for (auto& compId : tag.buildingTarget.componentIds)
+		if (tag.buildingTarget.progress < 100.0)
 		{
-			const auto ent = getWorld().getEntity(compId.first);
+			tag.buildingTarget.progress += duration * (1.0 / double(tag.buildingTarget.componentIds.size()));
+			
+			return;
+		}
 
-			if (!ent.isValid() || !ent.hasComponent<Item>())
+		auto skillCheck = skillRoll(e.getComponent<Stats>(), "Construction", DIFFICULTY_NORMAL);
+
+		if (skillCheck >= SUCCESS)
+		{
+			auto target = getBuilding(tag.buildingTarget.tag);
+			if (target == nullptr)
+				throw std::runtime_error("Building tag not found!");
+
+			auto& buildingEntity = getWorld().getEntity(tag.buildingTarget.entity_id);
+
+			auto & building = buildingEntity.getComponent<Building>();
+
+			// Check every component for validity
+			for (auto& compId : tag.buildingTarget.componentIds)
 			{
-				designations->buildings.push_back(tag.buildingTarget);
-				std::cout << "Invalid component for building - building id: " << tag.buildingTarget.entity_id 
-					      << " componnent id:" << compId.first << "\n";
-				work.cancel_work(e);
-				return;
+				const auto ent = getWorld().getEntity(compId.first);
+
+				if (!ent.isValid() || !ent.hasComponent<Item>())
+				{
+					designations->buildings.push_back(tag.buildingTarget);
+					std::cout << "Invalid component for building - building id: " << tag.buildingTarget.entity_id
+						<< " componnent id:" << compId.first << "\n";
+					work.cancel_work(e);
+					return;
+				}
 			}
+
+			// Set materials for building and delete componenets 
+			for (auto& compId : tag.buildingTarget.componentIds)
+			{
+				const auto comp = getWorld().getEntity(compId.first);
+
+				const std::string comptag = comp.getComponent<Item>().tag;
+				std::size_t material = comp.getComponent<Item>().material;
+
+				itemHelper.deleteItem(compId.first);
+
+				building.materials.push_back(std::make_pair(comptag, material));
+			}
+
+			building.complete = true;
+
+			// Just used for filtering
+			buildingEntity.addComponent<RenderComponent>();
+
+			buildingEntity.activate();
+
+			// Update availible reactions for buildings if this is a new Building type
+			defInfo->updateBuildingReactions(building.tag);
+
+			// Add code for building provides once added in 
+
+
+			work.cancel_work(e);
+			return;
 		}
-	
-		// Set materials for building and delete componenets 
-		for (auto& compId : tag.buildingTarget.componentIds)
-		{
-			const auto comp = getWorld().getEntity(compId.first);
 
-			const std::string comptag = comp.getComponent<Item>().tag;
-			std::size_t material = comp.getComponent<Item>().material;
+		else if (skillCheck == FAIL)
+			tag.buildingTarget.progress = 50.0;
+		
+		else
+			tag.buildingTarget.progress = 0.0;
 
-			itemHelper.deleteItem(compId.first);
-
-			building.materials.push_back(std::make_pair(comptag, material));
-		}
-
-		building.complete = true;
-
-		// Just used for filtering
-		buildingEntity.addComponent<RenderComponent>();
-
-		buildingEntity.activate();
-
-		// Update availible reactions for buildings if this is a new Building type
-		defInfo->updateBuildingReactions(building.tag);
-
-		// Add code for building provides once added in 
-
-
-		work.cancel_work(e);
+		return;
 	}
 }
